Rename `testutil-github` to `testutil-mentions`.
