+++ mentions/src/test/kotlin/io/spine/examples/pingh/mentions/given/GitHubClientServiceSpecEnv.kt
@@ -37,7 +37,7 @@ import io.spine.examples.pingh.github.NodeId
 import io.spine.examples.pingh.github.PersonalAccessToken
 import io.spine.examples.pingh.github.User
 import io.spine.examples.pingh.github.buildBy
-import io.spine.examples.pingh.testing.github.given.PredefinedGitHubResponses
+import io.spine.examples.pingh.testing.mentions.given.PredefinedGitHubResponses
 import io.spine.net.Url
 
 /**
